demo(android): unify custom-mesh + shape into custom-geometry (#2239 Batch 1)
Consolidates the retired `custom-mesh` and `shape` demos into a single
`custom-geometry` entry with a top segmented-button toggle between two
modes:

- **Custom Mesh** — composite primitives (SphereNode + CylinderNode
  molecule) with auto-rotate + scale slider. Default landing mode.
- **Shape Extrude** — 2D polygon paths triangulated into 3D meshes
  (Triangle / Star / Hexagon chips).

The per-mode controls + camera framing are preserved verbatim — this is
pure UI consolidation, no logic rewrite.

### Deep-link aliases

The retired `custom-mesh` and `shape` ids stay routable via
`DeepLinkRouter.DEMO_ID_ALIASES`. Existing `sceneview://demo/custom-mesh`
and `sceneview://demo/shape` deep links continue to land on the unified
demo (with the default Custom Mesh tab — tab pre-selection from alias is
a planned follow-up; not in Batch 1 scope).

Maestro flows (`.maestro/android/advanced.yaml`) reference both retired
ids and only assert the per-demo crash-gate ("Navigate back" visible) —
they keep passing via the alias without any flow edits.

### Tests

- `DemoInteractionTest.customMesh_autoRotateAndOrbit` — opens the
  unified demo (default Custom Mesh tab) and exercises auto-rotate /
  orbit / scale slider as before.
- `DemoInteractionTest.shape_allPolygons` — opens the unified demo,
  taps "Shape Extrude" to switch tabs, then cycles Triangle / Star /
  Hexagon.
- `DemoRenderingScreenshotTest.customGeometryDemo_default_state` —
  renamed from `customMeshDemo_default_state`, points at the new
  `custom-geometry` slug. Stale `custommesh_default.png` and
  `shape_default.png` goldens removed; the harness will re-capture on
  first run.

### Quality gates

- ✅ `./gradlew :samples:android-demo:compileDebugKotlin`
- ✅ `./gradlew :samples:android-demo:assembleDebug`
- ✅ `./gradlew :sceneview:detekt :arsceneview:detekt :sceneview-core:detekt`
- ✅ `bash samples/android-demo/scripts/collate-demos.sh --check`

### Out of scope

- iOS mirror — Android-only per Batch 1 plan; iOS follows in a later
  sprint (`feedback_always_all_platforms.md` is preserved as a TODO in
  the issue, not a blocker for this PR).
- Tab pre-selection from alias — follow-up, will let
  `sceneview://demo/shape` land on the Shape Extrude tab automatically.

Refs #2239 Batch 1 of 4.
