--- samples/android-demo/src/main/res/values/strings_demo_custom_mesh.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `custom-mesh` demo (registered by
-     `CustomMeshFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_custom_mesh_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_custom_mesh_title">Custom Mesh</string>
-    <string name="demo_custom_mesh_subtitle">Custom vertex and index buffers</string>
-</resources>
