--- samples/android-demo/src/main/res/values/strings_demo_shape.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `shape` demo (registered by
-     `ShapeFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_shape_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_shape_title">Shape Extrude</string>
-    <string name="demo_shape_subtitle">Extrude 2D polygons into 3D meshes</string>
-</resources>
